Correctly dock first view in workspace
This turned out to be a programmer error.

It was pointed out that docking the initial workspace should be done in the onBeforeShow method of the application path, not in the initialization of the workspace.

The program works now with the current head (11a36316) of tornadofx.

The seemingly erroneous warning about the invisible workspace remains, however.